feat(maps): add individual reset setting to default button (#39)
Adds a "Reset to Default" icon button next to any map setting that has been changed from the default setting. Also remove the old "Reset to Default" button that would reset all settings to default and replaces it with a "Save" button for an easier way to quickly exit the modal and save your changes.

Closes [CODZG-165](https://linear.app/codzombiesguides/issue/CODZG-165/add-save-button-to-map-setting-panel)