feat: generate image paths for full type-safety (#63)
Adds a `generate:image-paths` script that will generate union types for every image path in every directory, excluding the `content` and `opengraph-images` directory.

We don't include the `content` directory simply because it is too large; there are over 2k images in that directory, and even though the file size of the type generation is irrelevant for the production build, that many string unions cause issues in the past with the TypeScript language server.

With `tsgo`, this is likely a solved issue, but it isn't as critical for guide content since all guides are thoroughly previewed and usually iterated on multiple times before shipping, unlike the other config-based content.